Remove Runner
The Runner object is more harm than it's worth.  Now just run everything in a pool, with a dictionary of arguments fed to each instance.  Each instance writes to the DB (which will make conversion to Postgres easier in the long run), and if that fails out just dumps to the output list.